refactor: remove dataset_examples submodule
